Make resetting of lexer in command loop conditional on all input being consumed
This change was prompted by the fact that a call to the GNU |readline| function
can return a string with embedded newline characters (the one at the end is
stripped) if the user did a multi-line paste during the call (when "bracketed
paste mode", a feature more or less recently added to readline, is enabled). The
logic of the main command loop was assuming that, since a command can only be
executed when a newline just has been processed, no further input remains in the
buffer so resetting it (which also prepares a new line to be processed) can be
done without losing input. This resetting is now made conditional on actually
being at the end of the input from the last call of |readline|.

The current solution is not yet ideal, as preparing for a new line of input
should really be a separate functionality of the lexer than resetting. Also the
mechanism for pin pointing errors should be made aware of the possibility of
embedded newlines in the input buffer.