Correct previous broken and incomplete implementation of |FPP_numers|
There was confusion about the size of the result, about the direction in which
the Weyl group elements should be multiplied, and about the nesting order of the
traversal of the lists, and no effort was made to correctly find the images of a
point ending up on the border of the fundamental parallelepiped. The hopefully
correct approach followed now, is to first flag on the Weyl group which elements
$w$, when applied to the fundamental alcove, result in an alcove allowing a
root-lattice shift into the FPP (this depends only on the left descent set of
$w$, so can be easily tabulated) together with that shift itself; then for each
$W$-image of a facet (barycentre), find the (possibly empty) set of generators
for which the image is on the simple reflection wall, and for all those move to
both sides of that wall (for 2^m possibilities if the set has m elements).
Moving off those walls gives us a proper left descent set which we look up, and
if appropriate (admits a root-lattice shift) shift our image weight accordingly.

More errors were found and corrected later, mostly just implementation error,
such as forgetting that the vector |states| of intermediate states needs a
sentinel to ensure that (1) |states.back()| is always defined and (2) when the
outermost iterator is advanced, an initial state remains in |states| to base the
new intermediate state upon. This sentinel at position |0| introduces an index
misalignment between |states| and |coset_lists|, with numerous possibilities for
coding errors, which were extensively explored before being set right.