Fix bug in |center_classifier::shifts|: change borrow into carry to integral part
When negative contribution to shift has larger fractional part at some root than
the fixed (positive) contribution, we must make the difference positive by
adding 1 so that it can match the fractional part of the (to be selected)
positive contribution. This used to be interpreted as borrowing from the
integral part into the fractional part but in fact by making fixed+positive one
more time the root than negative, the fractional part will actually carry one
times that root into the integral part.