Extend loop syntax to allow nested conditionals and loops producing single list
This is a syntactic sugar change: some new syntax is expanded in the parser to
nested loops and conditionals with the appropriate calls of the concatenation
operator ##@([[T]]) to flatten the result to a simple list. For example

   for i:9 if i%3!=0 do i fi od

translates to

   ##(for i:9 do if i%3!=0 then [i] else [] fi od)

and produces the list [1,2,4,5,7,8]. Similarly a nested for loop

   for i:4 for j:i do (i,j) od od

translates to

   ##(for i:4 do for j:i do (i,j) od od)

and produces the list [(1,0),(2,0),(2,1),(3,0),(3,1),(3,2)]. Nesting can go on
to any depth, for instance to generate relatively prime pairs one can do

   for i:10 for j:10 if gcd(i,j)=1 do (i,j) fi od od

which translates to

   ##(for i:10 do ##(for j:10 do if gcd(i,j)=1 then [(i,j)] else [] fi od) od)

and produces a single list of length 57. The new syntax is triggered by using a
nested |if| or |for| instead of |do|; the header clauses end when an actual |do|
is seen, and at the end every header has its own closing |fi| or |od| (the
latter makes opening and closing tokens balanced in all cases, which is
important for the lexical analyzer which allows a newline character to signal
the end of a command only when such a balance has been reached). In fact the
outer construct might start with |if| rather than |for|, and is distinguished by
the fact that there is no |then| or |else|, but |do| (or a nested loop) instead.