Made facetious use MvL's new hash functions
Typical new command is local@(KGBElt x, ratvec lambda). This is meant
to find all unitary reps with final standard parameter (x, lambda, nu)
and with infl char in the FPP. I have tested it a little for the equal
rank case, and a very little in unequal rank. Doesn't seem to fail,
but there have been little issues: the new command seems to mess up
for complex G.