Mark |install_function| calls for functions where new optimisation could apply
Basically these are built-in functions that return a single value, and either
have a single argument which is of that type, or two arguments (at least) one of
which is of that type. It is not essential that the function try to |get_own|
the argument in question, even though often it does. Indeed, some functions that
could benefit from using |get_own| but did not were changed to henceforth do so.