Fix bug in |FPP_w_shifts|: wrong order of traversal of reflection list
In spite of all my thinking about this, the proper order to process the
reflections returned by the internal |to_positive_system| in order to change a
Weyl group element w to one that will make w*gamma integrally dominant with the
same integrality datum is the original, natural one. So |to_positive_system| was
changed to produce this natural order; and |wrt_distinguished| which needs the
opposite order just has to suck it up and reverse the list explicitly.