Improve integrality handling: store data at fundamental alcove; use W action
This fairly large reorganisation involved restructuring |integral_datum_item| (a
class defined in subsystem.h) to store a |SubSystem| that is assumed to be
generated by coroots associated to certain walls of the fundamental alcove (so
they are either simple coroots of the most negative coroot), and provide methods
to generate |codec| values for it that are transformed by an element |w| of |W|;
the assumption, which the code respects, is that |w| is chosen so as to map
simply-integral roots to positive roots. In the course of the revision, we
needed to modify |common_block| to avoid storing any |SubSystem|: the one used
during construction is made to be a local variable of the constructor, and
methods like |fold_orbits| that depended on the |SubSystem| reference that used
to be part of |common_clock| now are rewritten to achieve their goal differently.