Have |subsystem::integral_datum_item::codec| constructor better use its argument
We pass the matrix whose rows are the simple coroots for the integral system
from the class |integral_datum_item| to the constructor (which does not have
access to an instance of the containing class), but nonetheless that constructor
actually managed to locate the claaing instance of |integral_datum_item| and
gets the matrix from there instead. Better use the matrix provided as argument.