Partial cleanup of lusztig_cells...
Lots of the file is junk that should probably be removed. What's
interesting is mostly around Sommers_dual and keep_trying

Sommers_dual@(OG, OM) takes nilpotent and distinguished OM
representing component, returns Sommers dual OGvee. I think this is
fine.

Achar_dual_image(OGvee) returns the list of components OEvee for OGvee
so that (OGvee, OEvee) is in the image of Achar duality. Have not
actually computed Achar duality, but these functions seem OK.

sigma_lusztig_keep_trying(stG, OG) returns a list with terms
(OM, [int], int)
Here OM represents a conj class x in A(OG); [int] is the corresponding
list of W reps sigma(O,x); and int is the cheating required in some
truncated induction to get this to work. The list [int] seems always
to have exactly one element; if we knew this for certain, would be
better just to return that int.

If OG is special, these sigma(O,x) seem to be exactly Lusztig's
reps. If OG is NOT special, then they are strangely behaved:
sigma(OG,1) is then NOT the Springer rep of OG, but seems instead to
be the corresponding special W rep. I forget whether I decided this
was a good thing from a representation-theoretic perspective.

There are functions like sigma_lusztig_simple(stG,OG) which are
the same as keep_trying, except that they truncate induction
in exactly the natural degree. The result is that the W-rep lists
[int] attached to a component can be empty: example is

G=Sp(6), OG = complex_nilpotent(G,[1,1,0]), OM <-> nonid component.

This example is special. It's undoubtedly not a coincidence that
(OG,OM) is NOT in the image of Achar duality; but I don't understand
the connection.