Fix final bug in integrality reform, had forgotten one more word reversal
The function |from_fundamental_alcove| did not take into account that Weyl words
are interpreted from right to left, and therefore was missing a reversal. This
did not affect the correctness of the computation globally, but did cause
integral systems that are not associated to the fundamental alcove to be stored.

It happens that the unique call of |from_fundamental_alcove| quickly reverses
the result and then ultimately reverses it back again; we resisted the
temptation to rename the function |to_fundamental_alcove|, avoiding reversal
inside and the skip the first reversal in the caller; while marginally more
efficient it would make it even harder to understand why no reversal was called
for inside the function (namely 4 reasons to reverse canceling out each other).