Introduce |block_modifier| struct, from which currently only |shift| is used
The purpose of the struct is to record the relation between a stored block and a
block of current interest that is isomorphic to it, so that parameters for
elements of the latter can be reconstructed from those of the former. While
currently the relation just involves a shift orthogonal to the integral coroots,
it is envisioned, when fewer blocks get to be stored, that the relation will
also involve the action of a Weyl group element that changes the integral system
but maps positive integral coroots to positive ones, and consequently a
permutation of the labels of the simply-integral reflections used in the block.