Adapt clients of |lookup_full_block| to its modification of return values
We also split built-ins |print_block| and |print_partial_block| into a version
printing a freshly generated (partial) block and one that relies on looking up
the block int the |Rep_table| for the real form; the latter has restricted
possibilities for the integral system, so may look a bit different; the Weyl
group element that needs to be applied, and if non trivial the permutation of
the integrally-simple reflections, are specified in the output.

In the current state, the built-in functions using |lookup_full_block| are all
adapted to its change; they are |print_common_block|, |block@Param|,
|block_Hasse| (which function now also returns the block itself), |KL_block|,
|dual_KL_block|, |W_graph|, and |W_cells|. The last two functions adapt their
tau-invariants to the reordering of the simply-integral reflections with respect
to the stored block they use, avoiding an inconsistency with integrality_datum.