Modify storage of blocks to record an original block with modifier data
Here the $W$ modification of the infinitesimal character to be associated to a
facet of the fundamental alcove is only done for the purpose of grouping blocks
with a common |Reduced_param|, but no actual block at the normalised
infinitesimal character is built (since it may not be compatible with the fixed
distinguished involution). Instead data is stored together with the block that
records its position with respect to this normalised infinitesimal character.

This commit compiles, but is not yet functional.