Move checks of structure of sub-blocks somewhat outward in the code
These checks were buried in the innards of |swallow| methods, obliging those to
pass the relevant permutation of the simple roots around everywhere. They have
now been moved slightly outward, but not completely, because the test for
extended blocks can only be called when and where extended blocks are found, and
this discovery happens inside the |swallow| methods. What we have achieved is
that at least the block structure test is not inside the KL table methods.