Redefine where the |shift| and |delta| arguments to the |ext_block| ctor go.
Since we no longer (systematically) shift a block before calling the |ext_block|
constructor, it should be made to behave as if such a shift were applied. While
it was already passed to the selection of elements to be considered for the
extended block, it may also effect the behaviour of |tune_signs|; therefore we
pass it there as well. On the other hand, since we are passing values from the
constructor explicitly, we no longer need to store |delta| in the |ext_block|.