Work on twisted block construction; now amputating extended blocks on swallow
The main positive work is to avoid shifting whole blocks, moving the shifting
operations inside when individual representatives are gathered (and then
transformed to a different attitude).

The code that tried to reconstruct extended blocks in a large block when their
smaller block is swallowed is now disabled, as it seemed impossible to steer
clear of all the interdictions that it should respect; the effect is that if the
same or a similar extended block and/or its KL data is needed later, these will
have to be recomputed. We can now do |GL(6,R).trivial.twisted_full_deform|
again, but are still having trouble with |GL(6,R).trivial.is_unitary|.