Provide extra fields in structures for future changes, not using them yet
We provide a |w| field for the attitude in blocks, but still work with the
actual integral system so that it remains set to the identity and not used. In
the subsystem compilation unit we provide |integral_datum_item| with methods
that will enable handling of a given attitude, but the remain unused too.