Make |Reduced_param| constructor compute attitude; prepare for making it faster
Adding |int_sys_nr| and |w| fields is a key to reducing the number of blocks,
but computing it for every parameter is time consuming, as this commit shows. We
therefore already prepare a constructor that accepts precomputed values for
these fields, although the current code does not yet call these constructors.