Fix a conceptual error: cofolded FA facet does not determine FA facet
The goal here is really to determine the integral system for a delta-fixed point
of the fundamental alcove. Knowing the set of cofolded coroots with integral
evaluation at the point (which were available through |on_wall_subset|) simply
does not determine the set of unfolded coroots with the same property, due to
the fact type 3 folding produces a folded coroot whose restriction to
delta-fixed vectors is twice the restriction of one of the unfolded coroots.
Therefore we removed and avoided the method |InnerClass::unfold_FA_facet|.