Remove |int_sys_nr| and |bm| fields from |common_context|
The presence of these fields forced setting them whenever a block is
constructed, but this limits infinitesimal characters to being delta-fixed. Now
the constructor that directly takes an infinitesimal character as argument no
longer does the computation, and therefore no longer imposes the restriction.

We changed types form |RootNbrList| (a vector) to |sl_list<RootNbr>| in various
(mostly recent) places, in order to facilitate implementation of this change.