Use |root_vertex_of_alcove| for finding (cofolded) attitude for $\gamma$
This is a major strategy change in the task of finding an element of $W^\delta$
such that the integral system at |gamma| is the image by it of an integral
system associated to the fundamental alcove (at a $\delta$-fixed point). We no
longer look just at wall orientations (which will not distinguish alcoves that
are translated by sums of fundamental weights), but goes for the component in
$W$ of the unique affine Weyl group element that sends the alcove associated to
$\gamma$ to the fundamental alcove. The function |root_vertex_of_alcove| is
useful in getting rid of the pure (root) translation part of the affine Weyl
group element, so that |factor_dominant| can then be used to find |w|. This
still has to be replaced by a minimal coset representative, as before, in this
case for the stabiliser of the coset $\gamma+R$ by the root lattice $R$.

The reason this is an essential improvement is that sums of fundamental weights
for the cofolded root datum are not always sums of fundamental weights for the
original root datum, but the corresponding statement for sums of root is true.
Thus a bug is avoided, in which the shift involved changes the integral
subsystem of the original root datum (but not of the cofolded one).