Correct omission: forgot to adapt Dynkin diagram of extended block to permutation
When an extended block is generated for a given block transformed by a cofolded
Weyl group element and shift, the order of the (cofolded) integrally simple
generators is changed (to match the order that would have been chosen if a block
had been present that needed no transformation), and the internals of the
extended block are adapted to this modification. The Dynkin diagram must be too,
so that the |check_braid| method will make the correct tests.