Significant reworking of FPP_, especially FPP_local.  Changes were meant to speed up FPP_unitary_hash2 and FPP_unitary_hash_one_level, but they have not yet succeeded in doing that. Hope this provides a setup for speed improvement.
There was a months-old really subtle bug, which I couldn't make appear
until some strange F4_s setting. In creating the "backward
inheritance" links for unitarity (where unitarity on a small facet S
implies unitarity on a big facet B having S in its closure, I
previously checked only that the K-char formulas on B and S were the
same. This allowed the possibility that S had two UNITARY Langlands
quotients LS and LS', and B had a SINGLE Langlands quotient LB, having
the same K-char as the sum LS + LS'. In this situation, the code
incorrectly put a backward arrow from S to B, and deduced that LB had
to be unitary.

Either this error, or something close to it, appeared in
x = KGB(F4_s,154), lambda = FPP_lambdas(x)[13]. Fixed it by adding to
the FPPVertexHash an extra coordinate recording the number of
Langlands quotients.

Corrected the way the unitarity test counter works, so that it counts
actual instances of calling a signature-deformation function. This
means that numbers like "#tests# from the old output will be smaller
in the new output.