Script paramRep.at is an idea for computing special rep(Param)
There are scripts for computing the special rep (of the integral Weyl
group) attached to a Param, but they get painfully slow for rank maybe
7. This script computes the "T_{alpha,beta} orbit" of p. That
certainly gives a lot of constraints on the special rep; don't know
whether it _determines_ the special rep. On my laptop, this script
found the 3045 q's in the orbit of one discrete series, in about 3
minutes and using less than  a gig of memory. Asking for the GK
dimension of p used four gigs of memory in three minutes; I don't know
whether the computation will finish, but I doubt it.