Correct confusing definition of |product@(int,[mat])|, left to right is natural
In multiplying a list matrices, there is really no reason to arrange it from
right to left, as was being done. Nor is it right to suppose the matrices
square, although it remains a good precaution to specify a (square) size for in
case the list is empty. Just one use in the library was found, and corrected.