Correct error in dirac_cohomology for disconnected G
Salah Mehdi and Pavle Pandzic pointed out that the function
dirac_cohomology@Param (which is meant in the equal rank case to
return the highest weights of the representation of K_0 on the
Dirac cohomology of a unitary representation) was giving wrong
answers for G disconnected. I hope this is corrected. Thank you,
Salah and Pavle!