Optimise expressions like |a[i+:=1] #:= new_term| using new expression type
Here |a| for instance has type |[vec]|; other row-of types and other operations
that can be done in-place if the row entry is unshared are also on the roster.

The main difficulty is making the entry unshared at the point the operation is
invoked, and from the language point of view the difficulty is that whether it
can be applied depends on the resolution of the overloading of the operator, at
a point in the analysis where expression-rewriting is no longer easy (as it
requires picking apart expressions in |expression_base| based format). The
conversion of these "component transform expressions" was no doubt the hardest
implementation challenge in the language so far, as is cuts across the different
language processing phases in an unusual way.