Fix amazingly old error: |conform_types| stole required type on error.
The second (|required|) argument of |conform_types| is passed by non-constant
reference so that it might if necessary be specialised. The function was however
abusing this licence to change |required| when throwing an error (in case
nothing works), by _moving_ the type from it into the error object. This can
cause types in many places to be inadvertently set to undefined, for instance
the type (or a sub-type) of a global variable, if that happens to provide the
required type for some call of |conform_types|. This error could also affect
|resolve_overload|, since it ends up calling |conform_types|. The fix to the
error was simple, just call |required.copy()| when throwing the error.