Fix lifetime issue of |expr| values in component transform statements
The problem was that this new code builds some |expr| values outside the parsing
phase, namely as temporary rearrangements of the parts of the syntax tree during
the context sensitive analysis (type checking and conversion), in order to find
out the binding of (overloaded) operator names. This means the strategy of
storing a reference to |expr| values in |expr_error| objects (which was inspired
by the fact that a copy constructor for |expr| would be a lot of work to write
and was therefore not done) produces a problem of dangling references when an
error is reported in terms of such a temporary expression, as it is in cases
where the operator matching fails.

The solution of the problem consists of introducing a few |static| variables to
hold the temporary expressions that so occur. Since this is in a highly
recursive function |convert_expr|, the old value of these variable must be
temporarily moved to local variables and restored on successful completion, so
that callers do not observe destruction or modification of the values stored in
these static variables. In correcting this, a similar error was also fixed: when
constructing the temporary |expr| values, we are using parts of the parser
generated syntax tree (again because there is no copy constructor), but these
must be moved back into their original places so that later error messages
involving the part of the syntax tree we worked on will not find it amputated of
certain parts (that were moved into temporaries and then destructed).