Implement conversion of expressions |x+1| to |succ(x)|, |x+:=1| to |x:=succ(x)|
This is realised by using a special virtual function |build_call| that tries to
recognise calls where the second argument is '1' (an integer denotation).
Since in |x+:=1| is rewritten as |x:=x+1| before type checking and conversion,
this becomes |x:=succ(x)|; also the argument of |succ| will be used in-place.
A similar conversion of |x-1| to |pred(x)| is implemented as well.