(Again!) fixed "all remaining" bugs in FPP_unitary_hash_bottom_layer for disconnected and unequal rank G.
This was a nightmare, with somewhat nightmarish results. The corrected
code is substantially slower than FPP_unitary_hash for small G (factor
of three slower for GL(6,R). But it uses MANY fewer is_unitary tests,
and so less time in is_unitary; this makes the bottom layer version
a bit faster and substantially smaller for F4_s. I'll test some
bigger G after pushing.

An important feature is that it works with derived groups of simple
factors of Levi subgroups, doing signature calculations only if it's
never encountered the same simple derived group before. Of course this was
also the source of many bugs: the functions simple_factors and derived
(both??) can reorder the simple roots.