Two big bugs: FPP_global failed (given KGBElt x) to find all lambda so that (x,lambda) could contribute to FPP_unitary dual. Pretty sure this is fixed, by fairly simple changes to various FPP_lambda* functions.
Second bug was in bottom_layer: the LKdata constructed to index the
bottom layer failed to do that, in all kinds of subtle ways (related
to synthetic real forms). Effect was that some unitary reps were
reported as NONunitary, because the requirements on K-mults imposed by
the (bad) LKdata were wrong.

This was awful to diagnose and fix. I'm not at all convinced that it's
OK, although I have checked that various unequal rank and disconnected
G at least don't crash the program. I will start testing bigger
exceptional groups, and this is likely to unearth some bugs.