incorporate new methods for matching KGBElts/KTypes/Params between L and its derived simple factors.
Changes are in cohom_reductionND.at. Goal is to take a reductive L
(typically Levi in G), look at its simple factors S (each living on
same X^* as L and G) and their derived groups derS.

New methods, like derived_factor_KGBs2, have a 2 after name. They
are intended to be "compatible" with things like LKT... and so
to avoid the need for "twist" in bottom_layer.at. I think this
works (tested in F4_s), but it's very slow.