speed tweaks
When non-unitary facets are eliminated, they are added to the
information in unitarity_hash.

cohomological induction in the weakly good range is necessarily
irreducible or zero; simplified tests in that setting accordingly.

Still to do: get is_weakly_unipotent test for Param, and use it
to get more unitary reps.