Fixed bug in bottom_layer.at appearing only in unequal rank.
The bottom_layer code sometimes assumed that the parabolic given
by theta_stable_quasi_data(p) is not a Borel (so that a nontrivial
bottom layer can exist). In equal rank, this is true whenever
bottom layer is called; but in unequal rank, not necessarily.
This cause a failure with G=E6_s, more_flag:=true.