Fix error that was introduced in optimisation: local variable may be found empty.
Our optimisation to temporarily move values from local variables if a combined
operation-assignment wants to modify the value in-place (for which it has to be
unshared at that point) may produce a scenario where the local variable holds no
value when an error is thrown. This code that prepares a back trace therefore
has to be able to handle this situation without crashing, and the current fix
makes it prepared to do that. Hopefully this is the only scenario in which the
temporary absence of a value from a local variable can be noticed by the user.