Significant speed improvements; correct bugs for disconnected G
Added "know_lower" functionality to other algorithms as in
bottom_layer.

Avoid repeating computations for contragredients of known params,
or for twists by group automorphisms coming from Ad(G). This is
a very large speed improvement: approximately a factor of the
two-torsion in Z(G).

Code in derived_cover.at for lifting KGBElts to the universal cover of
the derived group fairly often failed: usual reason that there are
several strong real forms of this cover, and the code lifted to
different ones for different KGBElts. I don't know a good fix for
this, so used the same kludge as in cohom_reductionND.at: the KGBElts
in these different strong real forms are number consistently with
isomorphisms between them.