Squashed a (fairly old?) bug in function local_testK_hash_bottom_layer
The function (when more_flag is true) applies a bottom layer function
bottom_elim_factors_more to eliminate some candidate unitary
reps. That function needs to compute the K-character of its
argument. Sometimes the K-char is already computed, and it can be
passed as an optional last argument to the bottom_layer function.
Unfortunately I passed it in cases (#ps > 1) when it was NOT known.

There are some speed issues still to be addressed, but I have not yet
seen other errors.