speed/cleanup for bottom layer.
The bottom layer tests were inadvertently being RE-applied at the last
stage of local_test_GEO_hash_bottom_layer. Took that out, which
provides a small speed increase. There is still some change in the
last few days which is SLOWING the code a bit; maybe it has to do with
the ORDER in which bottom_layer and wiggle arguments are applied
somewhere??