misunderstood bottom_layer
In local_test_GEO_hash_bottom_layer, I was as usual getting
unitary-on-bottom-layer vertices and edges, then adding
higher-dimensional facets one dimension at a time, checking for bottom
layer unitarity at each dimension. This second step turns out (at
least up to rank 6???) to be a waste of time: we're looking at facets
all of whose boundaries are unitary on the bottom layer, so (almost
always in theory, and apparently _always_ in practice?) these new
facets are unitary on the bottom layer as well.

So now I just add the facets, omitting all the bottom layer checking,
and then do is_unitary tests on the resulting candidates at the end.
This eliminates quite a bit of code, but not much time: the bottom
layer stuff was organized to avoid repeating calculations, and this
apparently made these unnecessary cases very fast.