Make implicit conversions apply inside denotations (part of constant folding)
The motivation for this was to make it easier to recognise at compile time
certain constant operands that are promoted from integer to a rational number,
so that for instance |x>0| can be rewritten to |>x| when |x| is a rational
expression (in which case the |0| also gets implicitly converted to rational).

By introducing some evaluation during compile time, this change breaks an
organisational separation that was maintained so far between compilation units.
From that perspective this change posed some design questions, but the actual
implementation, notably of its general applicability to all implicit conversions
at once, was quite easy. This gives some hope that constant folding can in the
future be applied without too much pain to all common types of constant
expressions.