some corrections to older algorithms (hash, etc), some speed tweaks
There is a function is_unitary_to_hts_hash which was wrong; I finally
detected the mistake in Sp(8,R). Issue is that it tests a parameter p
for unitarity to height h_G by realizing p as weakly good range
induced from p_L; then testing p_L for unitarity to height

h_L = h_G - height(p) + height(p_L).

This is nearly correct but not quite; my guess is that it can only
fail in one direction (that p_L can fail but p_G pass???) and that
it would be OK if one insisted on GOOD range. But I'm sick of bug
chasing, so I just marked the function as incorrect and removed
all calls to it.

Added a flag "quick_flag," defaulting to true. When it's true, the
unitary candidate facets of dimensions 2 to top are taken just to
be those with unitary candidate boundaries: no further testing is
done on them until the end, when all candidates are tested by
is_unitary_hash. ("false" is the old behavior, that some tests are
performed on the new facets of dimensions 2 to top as they are found.)

This code seems to fix an error in FPP_unitary_hash that had been
present for a long time; and seems to improve the speed of everything
a small amount. I've tested it in quite a few examples up to rank 4,
and it seems good; after I push it, I will test E6_q and E6_s and try
to address any problem very quickly.