Make |expression_base::level| to a typedef for a scoped enumeration.
Extracting the actual definition of the enumeration to be outside the class
|expression_base| is necessary to enable the enumeration to be predeclared,
which is a prerequisite for being able to introduce |wrapper_function| as a
typedef and therefore to use typedefs that refer to |wrapper_function| in
predeclarations. Making it scoped (so that enumeration values need to be
qualified) compensates somewhat for the unwanted global exposure of the
enumeration constants that this move would otherwise entail.

The rules for declaring, defining and using enumerations in C++ are quite a
mess, which explains why this leads to quite a painful and pervasive change.