Use previous change to have |install_function| return a |shared_builtin| value
The type of this value is pointer to a type derived from |function_base|, where
pointer to |function_base| was previously used for a silly reason:
|function_base| is a non-template class, and the identifier could therefore be
predeclared, whereas |builtin| is a typedef for a template instance that is
derived from |function_base|, and as such cannot be predeclared. With the
reoganisation done, the file axis-types-fwd.h contains the declarations
necessary to define the return type of |install_function| where it is needed.

This change obviates the need to use |static_cast| to cast back down to the
|builtin| derived type, which facilitates installing specila tests.