Complete the marking of the set of built-ins that perform constant folding
In many cases it is questionable whether constant folding will every be useful
for a given operation (i.e., applied on an expression that is not immediately
executed a single time afterwards). We have tried to draw the line at operations
that can serve to build up or transform in a straightforward manner values,
excluding operations that do nontrivial computation (like |gcd|) or which lead
values smaller than their operands (row or column extraction from a matrix,
scalar product of vectors). As somewhat of an edge case we have included
arithmetic operations between matrices, vectors and rational vectors. Sometimes
constant folding will require more memory that the original expression would
(say for a large identity matrix) but if this really is undesired, there are
easy ways to block unwanted constant folding (like using a |let| expression).