speed tweaks
added a bit more to_ht testing to bottom_layer. The bad point [(20269,rho)]
for E7_s now runs in less than 200 seconds on my laptop.

There was a bug in isotwistKGBs in FPP_global which made it fail in
SU(3,2). Fixed with the usual kludge of getting a KGBElt _number_;
waiting for Jeff to find the solution from The Book of this problem.