speed tweaks to FPP_local
The issue of when to use is_unitary_to_ht is a real pain: for the rep

p = parameter(E7_s, 20269, E7_s.rho,[8,0,0,0,0,7,2])

is_unitary takes hours, but is_unitary_to_ht(p,114) returns "false"
in a few seconds. But there are other examples with essentially the
opposite behavior: is_unitary_to_ht(q,[small height]) takes a very
long time, but is_unitary returns false very quickly.

This version of the script is better than the last, but still needs
a lot of work. "bottom_layer" is pretty good. "hash2" is OK by itself,
but when it reverts to hash_dumb (as it does for the pair (20269,rho)
above) there is trouble.

setting "revert_flag:=false" prevents this bad reversion, so might
be a good choice with hash2 for now.

I am pushing this partly to test it on the spherical dual of E8_s.
Old code took four days/280g. We'll see if this at least matches that.