speed tweaks...
rewrote theta_induced@RealForm to be parallel to real_induced. In
this version it's much faster (as expected) and finds a lot more
cohomologically induced reps (which I don't understand).

This is getting to be in reasonable shape. A good thing to add
would be records of known non-unitary facets, obtained by real
or cohomological induction. This involves Lucas's chop_facets.