Add interrupt checking to recursive |full_deform| built-in function
Though a small change, this requires some reorganisation, as it is the first
time that a library function calls a function defined in the interpreter. To
avoid having to include interpreter header files from any library files (in the
current case from repr.cpp), the (very simple) declaration of the function
|interpreter::check_interrupt| was added to Atlas.h (making the one that appears
in interpreter/axis.h redundant) so that anyone can safely call this function.

This also requires a dummy definition of |interpreter::check_interrupt| to be
added (in a new file test/interrupt.h) to enable linking of the Fokko program,
since it uses repr.o which object file will refer to |check_interrupt|.