fixed bug in FPP_unitary_hash_def_dumb
Incorrectly used global vertex data instead of local vertex data. Can
affect other algorithms because of "reversion"; the error was obscured
because usually reversion happens close to the spherical case, when
the local vertices are close to the global vertices.

Renamed a bunch of arguments "Lvd" when they needed to be local.