Added ability to precompute nonunitarily induced representations
Idea is to look at maximal real Levis in G. For each, look at all
known non-unitary hermitian reps on each Levi. If the induced rep to G
is irreducible and FPP, then label it as nonunitary.

For the moment this computation is quite slow, and the knowledge of
these nonunitary reps on G seems not to help the FPP_unitary
calculation on G very much.