More modifications to is_unitary_to_ht...
...which seem to address speed problems still present in the previous
commit. For example in the previous commit

FPP_unitary_hash2([(KGB(E7_s, 5366),[  2,  1, -1,  2,  1,  1,  0]/1)])

took more than fifteen minutes on my laptop; now it's about 13 seconds.

The idea is that each is_unitary_to_ht test looks at the relative
sizes of lambda and nu. If lambda is bigger, it computes the full
signature and truncates to height. If nu is bigger, then it uses
Marc's to_height calculation (more or less in the dual KL block).
This is implemented at the end of FPP_facets_herm.at; changes to
FPP_localDirac.at ensure that the new functions are actually called.

The rational parameter low_frac defaults to 1. If it is assigned to 0,
then Marc's algorithm is almost always used. If it is assigned a large
value, then usually one truncates the full signature.